requirements.txt
absl-py==0.10.0
altgraph==0.17
appdirs==1.4.4
apturl==0.5.2
argcomplete==1.10.0
argon2-cffi==20.1.0
asn1crypto==0.24.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
audioread==2.1.8
auto-py-to-exe==2.7.8
autocorrect==0.4.4
backcall==0.2.0
beautifulsoup4==4.8.0
bleach==1.5.0
blis==0.7.4
boto==2.49.0
boto3==1.14.18
botocore==1.17.18
bottle==0.12.18
bottle-websocket==0.2.9
Brlapi==0.6.6
cachetools==4.1.1
catalogue==1.0.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.3.7
command-not-found==0.3
contextvars==2.4
contractions==0.0.21
cryptography==3.4.7
cssselect==1.1.0
cssutils==2.3.0
cupshelpers==1.0
cycler==0.10.0
cymem==2.0.5
Cython==0.21
dataclasses==0.8
datefinder==0.7.1
dateparser==0.7.6
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
devscripts===2.17.12ubuntu1.1
DeWave==0.12
distlib==0.3.1
distro-info===0.18ubuntu0.18.04.1
docutils==0.15.2
docx2txt==0.8
EbookLib==0.17.1
Eel==0.12.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints==0.3
enum34==1.1.10
extract-msg==0.23.1
filelock==3.0.12
Flask==2.0.1
Flask-PyMongo==2.3.0
future==0.18.2
gast==0.4.0
gensim==3.4.0
gevent==20.6.2
gevent-websocket==0.10.1
gluoncv==0.8.0
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
gpg==1.10.0
graphviz==0.8.4
greenlet==0.4.16
grpcio==1.26.0
h5py==2.10.0
html5lib==0.9999999
httplib2==0.9.2
idna==2.10
imageio==2.9.0
IMAPClient==2.1.0
imgviz==1.2.2
immutables==0.15
importlib-metadata==4.6.1
importlib-resources==3.0.0
imutils==0.5.3
inflect==2.1.0
invoice2data==0.3.5
InvoiceNet @ file:///home/diya/Desktop/InvoiceNet
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kazam==1.4.5
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-resnet==0.2.0
keras-retinanet==0.3.1
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
labelImg==1.8.3
labelme==4.5.6
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
librosa==0.6.2
llvmlite==0.31.0
louis==3.5.0
lxml==4.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.3.1
MarkupSafe==2.0.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==8.5.0
murmurhash==1.0.5
mxnet==1.7.0.post1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
networkx==2.4
nltk==3.6.2
nose==1.3.7
notebook==6.4.0
notify2==0.3
numba==0.43.0
numpy==1.19.5
oauth==1.0.1
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==3.4.2.16
opencv-python==4.2.0.32
opt-einsum==3.3.0
packaging==21.0
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
pathy==0.6.0
pdf2image==1.13.1
pdfminer==20191125
pdfminer.six==20201018
pdfplumber==0.5.22
pexpect==4.2.1
pickleshare==0.7.5
Pillow==8.3.0
plac==1.1.3
pluggy==0.6.0
portalocker==2.0.0
premailer==3.9.0
preshed==3.0.5
progressbar2==3.53.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.13.0
ptyprocess==0.7.0
py==1.5.2
pyahocorasick==1.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycairo==1.16.2
pycocotools==2.0.1
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.10.1
pycups==1.9.73
pydantic==1.7.4
pydicom==2.1.2
pydub==0.23.1
Pygments==2.9.0
PyGObject==3.26.1
pyinstaller @ https://github.com/pyinstaller/pyinstaller/archive/develop.zip
pyinstaller-hooks-contrib==2020.7
pymacaroons==0.13.0
pymongo==3.11.0
PyNaCl==1.1.2
pyparsing==2.4.7
PyPDF2==1.26.0
PyQt5==5.15.1
PyQt5-sip==12.8.1
pyresparser==1.0.6
pyRFC3339==1.0
pyrsistent==0.18.0
pytesseract==0.3.4
pytest==3.3.2
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
python-debian==0.1.32
python-magic==0.4.16
python-pptx==0.6.18
python-utils==2.4.0
python3-protobuf==2.5.0
pyttsx3==2.81
pytz==2021.1
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.3.1
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.9.0
regex==2021.7.6
reportlab==3.4.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
resampy==0.2.2
rsa==4.6
s3transfer==0.3.3
scikit-image==0.15.0
scikit-learn==0.23.2
scikit-video==1.1.11
scipy==1.5.4
SecretStorage==2.3.1
Send2Trash==1.7.1
simplegeneric==0.8.1
simplejson==3.17.2
six==1.16.0
sklearn==0.0
smart-open==5.1.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
soupsieve==2.0.1
spacy==2.3.5
spacy-legacy==3.0.6
SpeechRecognition==3.8.1
srsly==1.0.5
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tensorboard==2.1.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.1.3
tensorflow-estimator==2.1.0
tensorflow-gpu==1.3.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
terminado==0.10.1
tesserocr==2.5.0
testpath==0.5.0
textract==1.6.3
textsearch==0.0.17
Theano==1.0.5
thinc==7.4.5
threadpoolctl==2.1.0
tifffile==2020.7.24
torch==1.2.0
torchvision==0.4.0
tornado==6.1
tqdm==4.61.1
traitlets==4.3.3
typer==0.3.2
typing-extensions==3.10.0.0
tzlocal==1.5.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unicodecsv==0.14.1
Unidecode==1.1.1
unidiff==0.5.4
urllib3==1.26.6
usb-creator==0.3.3
virtualenv==20.0.34
wadllib==1.3.2
Wand==0.6.2
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5
Werkzeug==2.0.1
whichcraft==0.6.1
widgetsnbextension==3.5.1
wordcloud==1.7.0
wrapt==1.11.2
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==1.2.9
zipp==3.5.0
zope.event==4.4
zope.interface==4.3.2
